:construction: Add draw_ray function
*
TODO:
- fix for S and W dir

+++ inc/structs.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   structs.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
+/*   By: foctavia <foctavia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 09:16:02 by owalsh            #+#    #+#             */
-/*   Updated: 2022/11/17 16:57:27 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/11/18 14:47:35 by foctavia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,7 +95,7 @@ typedef struct s_camera
 
 typedef struct s_player
 {
-	int			dir;
+	float		dir;
 	float		angle;
 	t_coord		pos;
 }				t_player;
